separating auth routes into separate files for cleaner structure and testing
