+++ app/templates/share.html
@@ -115,7 +115,7 @@
     </div>
     <!-- New fixed logout button (bottom-right) -->
     <div class="fixed bottom-4 right-4">
-        <a href="{{ url_for('main.logout') }}"
+        <a href="{{ url_for('auth.logout') }}"
             class="bg-red-500 hover:bg-red-600 text-white font-bold py-2 px-4 rounded-full shadow-lg transition duration-200 flex items-center">
             <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5 mr-2" viewBox="0 0 20 20" fill="currentColor">
                 <path fill-rule="evenodd"
