+++ app/templates/welcome.html
@@ -81,13 +81,13 @@
       </h2>
 
       <div class="flex justify-center space-x-4">
-        <a href="{{ url_for('main.signup') }}">
+        <a href="{{ url_for('auth.signup') }}">
           <button
             class="bg-button text-secondary text-xl font-bold py-4 px-8 rounded-full hover:bg-opacity-90 transition duration-200 shadow-md">
             Sign Up
           </button>
         </a>
-        <a href="{{ url_for('main.login') }}">
+        <a href="{{ url_for('auth.login') }}">
           <button
             class="bg-button text-secondary text-xl font-bold py-4 px-8 rounded-full hover:bg-opacity-90 transition duration-200 shadow-md">
             Log In
